requirements.txt
plotly
ipyvizzu
pandas 
numpy
seaborn 
matplotlib
streamlit-aggrid